// frontend/src/api/generated/schemas/finishResultDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * onBoard API
 * AI-powered interview preparation service
 * OpenAPI spec version: 1.0
 */

export interface FinishResultDto {
  sessionId: string;
  status: string;
  totalQuestions: number;
  questionsAnswered: number;
  avgScore: number;
  sessionScore: number;
  newFullScore: number;
  newLeague: string;
}